requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy[asyncio]==2.0.36
psycopg2-binary==2.9.10
asyncpg==0.30.0
pydantic-settings==2.6.1
email-validator>=2.0.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
slowapi==0.1.9
prometheus-client==0.20.0
psutil==6.1.0
